requirements.txt
paho-mqtt~=2.1.0
pyotp~=2.9.0
smartrent-py~=0.5.2
structlog~=25.5.0
